backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pytube==15.0.0
redis==5.0.1
google-cloud-speech==2.21.0
google-cloud-translate==3.12.0
google-cloud-texttospeech==2.14.1
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1
yt-dlp==2025.02.19
ffmpeg-python==0.2.0
spacy==3.7.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
demucs==4.0.1
pydub==0.25.1
google-cloud-storage==2.14.0
pandas==2.2.0 